pyproject.toml
[tool.poetry]
name = "the-house"
version = "0.1.0"
description = ""
authors = ["zainaraza43 <urboyzain@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.4.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
